+++ app/controllers/AnnotationController.scala
@@ -2,7 +2,7 @@ package controllers
 
 import javax.inject.Inject
 import akka.util.Timeout
-import com.scalableminds.util.reactivemongo.{DBAccessContext, GlobalAccessContext}
+import com.scalableminds.util.accesscontext.{DBAccessContext, GlobalAccessContext}
 import com.scalableminds.util.tools.{Fox, FoxImplicits}
 import com.scalableminds.webknossos.datastore.tracings.TracingType
 import models.annotation._
