+++ app/models/annotation/handler/AnnotationInformationHandler.scala
@@ -1,6 +1,6 @@
 package models.annotation.handler
 
-import com.scalableminds.util.reactivemongo.DBAccessContext
+import com.scalableminds.util.accesscontext.DBAccessContext
 import com.scalableminds.util.tools.{Fox, FoxImplicits}
 import models.annotation.AnnotationTypeSQL.AnnotationTypeSQL
 import models.annotation._
