+++ app/models/annotation/handler/ProjectInformationHandler.scala
@@ -1,6 +1,6 @@
 package models.annotation.handler
 
-import com.scalableminds.util.reactivemongo.DBAccessContext
+import com.scalableminds.util.accesscontext.DBAccessContext
 import com.scalableminds.util.tools.{Fox, FoxImplicits}
 import models.annotation._
 import models.project.ProjectSQLDAO
