+++ app/models/annotation/handler/SavedTracingInformationHandler.scala
@@ -1,6 +1,6 @@
 package models.annotation.handler
 
-import com.scalableminds.util.reactivemongo.DBAccessContext
+import com.scalableminds.util.accesscontext.DBAccessContext
 import com.scalableminds.util.tools.TextUtils._
 import com.scalableminds.util.tools.{Fox, FoxImplicits}
 import models.annotation._
