+++ app/models/team/Organization.scala
@@ -1,6 +1,6 @@
 package models.team
 
-import com.scalableminds.util.reactivemongo.DBAccessContext
+import com.scalableminds.util.accesscontext.DBAccessContext
 import com.scalableminds.util.tools.{Fox, FoxImplicits}
 import com.scalableminds.webknossos.schema.Tables._
 import play.api.i18n.Messages
