+++ app/models/user/User.scala
@@ -2,7 +2,7 @@ package models.user
 
 import com.mohiva.play.silhouette.api.util.PasswordInfo
 import com.mohiva.play.silhouette.api.{Identity, LoginInfo}
-import com.scalableminds.util.reactivemongo._
+import com.scalableminds.util.accesscontext._
 import com.scalableminds.util.tools.{Fox, FoxImplicits, JsonHelper}
 import com.scalableminds.webknossos.schema.Tables._
 import reactivemongo.play.json.BSONFormats._
