+++ app/models/user/time/TimeSpan.scala
@@ -1,6 +1,6 @@
 package models.user.time
 
-import com.scalableminds.util.reactivemongo.DBAccessContext
+import com.scalableminds.util.accesscontext.DBAccessContext
 import com.scalableminds.util.tools.{Fox, FoxImplicits}
 import com.scalableminds.webknossos.schema.Tables._
 import models.annotation.AnnotationSQL
