+++ app/models/user/time/TimeSpanService.scala
@@ -5,7 +5,7 @@ package models.user.time
 
 import akka.actor.{Actor, Props}
 import com.scalableminds.util.mail.Send
-import com.scalableminds.util.reactivemongo.{DBAccessContext, GlobalAccessContext}
+import com.scalableminds.util.accesscontext.{DBAccessContext, GlobalAccessContext}
 import com.scalableminds.util.tools.{Fox, FoxImplicits}
 import com.typesafe.scalalogging.LazyLogging
 import models.annotation._
